6th Commit
Started over with the Sass file from scratch, but left my includes and
mixins. Changed the naming of the ids in the html. Also took my time
with each section so it styled properly.

Moved the search bar to the left above the sidebar. Also added in all
the months for the recent posts section. Need to add in a title above
the list. The search bar, menu, and recent posts lists are spaced out
from the post content area.